bug fixes, keypad work

+++ include/keypad.h
@@ -3,18 +3,18 @@
 
 #include <stdint.h>
 
-#define K0 (1 << 0)
-#define K1 (1 << 1)
-#define K2 (1 << 2)
-#define K3 (1 << 3)
-#define K4 (1 << 4)
-#define K5 (1 << 5)
-#define K6 (1 << 6)
-#define K7 (1 << 7)
-#define K8 (1 << 8)
-#define K9 (1 << 9)
-#define KS (1 << 10)
-#define KP (1 << 11)
+#define K0 (uint16_t)(1 << 0)
+#define K1 (uint16_t)(1 << 1)
+#define K2 (uint16_t)(1 << 2)
+#define K3 (uint16_t)(1 << 3)
+#define K4 (uint16_t)(1 << 4)
+#define K5 (uint16_t)(1 << 5)
+#define K6 (uint16_t)(1 << 6)
+#define K7 (uint16_t)(1 << 7)
+#define K8 (uint16_t)(1 << 8)
+#define K9 (uint16_t)(1 << 9)
+#define KS (uint16_t)(1 << 10)
+#define KP (uint16_t)(1 << 11)
 
 void keypad_init(void);
